+++ xtask/src/main.rs
@@ -552,7 +552,7 @@ fn run_ci_checks(workspace: &Path, args: CiArgs) -> Result<()> {
             build_documentation(
                 workspace,
                 BuildDocumentationArgs {
-                    packages: vec![Package::EspHal, Package::EspRadio, Package::EspHalEmbassy],
+                    packages: vec![Package::EspHal, Package::EspRadio, Package::EspPreempt],
                     chips: vec![args.chip],
                     ..Default::default()
                 },